ADDED: - full implementation & doxygen headers for hal_adc & sen_prox        - conquest_setLastNode(), conquest_getLastNode() and conquest_convertDirMaskToNode() CHANGED: - one RTC time unit is now 1ms (was 10ms)          - introduced HAL_MOTORS_FULL_TURN_STEPS constant and removed several other similar local defintions          - minor comment changes          - adjusted interrupt priorities; UART transmitter is now above RTC FIXED: hal_motors_init() returns wrong values

+++ trunk/e-puck/firmware/hal_adc.h
@@ -7,7 +7,7 @@
 
 void hal_adc_init( void);
 
-uint8_t hal_adc_read(
+uint16_t hal_adc_read(
 	IN const hal_adc_EChannel _eChannel
 	);
